JE 0.4
Resources are now integrated into the Resources folder and accessed using ResourceLoader.
This is a much better approach to getting resources as they are integrated within JAR builds.

Physics Fix: When dragging the window rendering pauses and the delta time increases while its held. This leads to systems that rely on delta time like physics to break and compensate for the time difference. There is now a hacky fix to prevent this. If the delta time is longer than 0.1 seconds (frame-rate is <10) delta time is set to 0.

Friend: Added get method if you want to use it to return a variable which only specific classes have access to.